Updated for Solr 6 and to use CloudSolrClient

+++ solrmeter/src/main/java/com/plugtree/solrmeter/model/statistic/FullQueryStatistic.java
@@ -137,7 +137,7 @@ public class FullQueryStatistic implements QueryStatistic {
 		double sumation = 0;
 		synchronized (sortedQueryTimes) {
 			for(Integer value:sortedQueryTimes) {
-				sumation+= Math.pow((double)(value - average), 2.0);
+				sumation+= Math.pow(value - average, 2.0);
 			}
 			return sumation / sortedQueryTimes.size();
 		}
